Implemented TVDB sync in Spring Batch, added initial sync and update sync jobs

+++ src/main/java/com/kwsni/caught_up/tvdb/dto/EpisodeBaseRecordDto.java
@@ -27,7 +27,7 @@ public record EpisodeBaseRecordDto (
     Integer runtime,
     Integer seasonNumber,
     List<SeasonBaseRecordDto> seasons,
-    Long seriesId,
+    Integer seriesId,
     String seasonName,
     String year
 ) {}
