+++ src/main/java/com/kwsni/caught_up/tvdb/repository/SeriesRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kwsni.caught_up.tvdb.model.Series;
 
-public interface SeriesRepository extends JpaRepository<Series, Long>{}
+public interface SeriesRepository extends JpaRepository<Series, Integer>{}
